+++ pybit/_v5_market.py
@@ -3,6 +3,19 @@ from .market import Market
 
 
 class MarketHTTP(_V5HTTPManager):
+    def get_server_time(self) -> dict:
+        """
+        Returns:
+            Request results as dictionary.
+
+        Additional information:
+            https://bybit-exchange.github.io/docs/v5/market/time
+        """
+        return self._submit_request(
+            method="GET",
+            path=f"{self.endpoint}{Market.GET_SERVER_TIME}",
+        )
+
     def get_kline(self, **kwargs) -> dict:
         """Query the kline data. Charts are returned in groups based on the requested interval.
 
@@ -264,5 +277,23 @@ class MarketHTTP(_V5HTTPManager):
         return self._submit_request(
             method="GET",
             path=f"{self.endpoint}{Market.GET_OPTION_DELIVERY_PRICE}",
+            query=kwargs,
+        )
+
+    def get_long_short_ratio(self, **kwargs):
+        """
+        Required args:
+            category (string): Product type. linear (USDT Perpetual only), inverse
+            symbol (string): Symbol name
+
+        Returns:
+            Request results as dictionary.
+
+        Additional information:
+            https://bybit-exchange.github.io/docs/v5/market/long-short-ratio
+        """
+        return self._submit_request(
+            method="GET",
+            path=f"{self.endpoint}{Market.GET_LONG_SHORT_RATIO}",
             query=kwargs,
         )