+++ pybit/_v5_spot_margin_trade.py
@@ -3,6 +3,20 @@ from .spot_margin_trade import SpotMarginTrade
 
 
 class SpotMarginTradeHTTP(_V5HTTPManager):
+    def spot_margin_trade_get_vip_margin_data(self, **kwargs):
+        """
+        Returns:
+            Request results as dictionary.
+
+        Additional information:
+            https://bybit-exchange.github.io/docs/v5/spot-margin-uta/vip-margin
+        """
+        return self._submit_request(
+            method="GET",
+            path=f"{self.endpoint}{SpotMarginTrade.VIP_MARGIN_DATA}",
+            query=kwargs,
+        )
+
     def spot_margin_trade_toggle_margin_trade(self, **kwargs):
         """UTA only. Turn spot margin trade on / off.
 
@@ -39,6 +53,34 @@ class SpotMarginTradeHTTP(_V5HTTPManager):
             path=f"{self.endpoint}{SpotMarginTrade.SET_LEVERAGE}",
             query=kwargs,
             auth=True,
+        )
+
+    def spot_margin_trade_get_status_and_leverage(self):
+        """
+        Returns:
+            Request results as dictionary.
+
+        Additional information:
+            https://bybit-exchange.github.io/docs/v5/spot-margin-uta/status
+        """
+        return self._submit_request(
+            method="GET",
+            path=f"{self.endpoint}{SpotMarginTrade.STATUS_AND_LEVERAGE}",
+            auth=True,
+        )
+
+    def spot_margin_trade_normal_get_vip_margin_data(self, **kwargs):
+        """
+        Returns:
+            Request results as dictionary.
+
+        Additional information:
+            https://bybit-exchange.github.io/docs/v5/spot-margin-normal/vip-margin
+        """
+        return self._submit_request(
+            method="GET",
+            path=f"{self.endpoint}{SpotMarginTrade.NORMAL_GET_MARGIN_COIN_INFO}",
+            query=kwargs,
         )
 
     def spot_margin_trade_normal_get_margin_coin_info(self, **kwargs):