+++ pybit/market.py
@@ -2,6 +2,7 @@ from enum import Enum
 
 
 class Market(str, Enum):
+    GET_SERVER_TIME = "/v5/market/time"
     GET_KLINE = "/v5/market/kline"
     GET_MARK_PRICE_KLINE = "/v5/market/mark-price-kline"
     GET_INDEX_PRICE_KLINE = "/v5/market/index-price-kline"
@@ -16,6 +17,7 @@ class Market(str, Enum):
     GET_INSURANCE = "/v5/market/insurance"
     GET_RISK_LIMIT = "/v5/market/risk-limit"
     GET_OPTION_DELIVERY_PRICE = "/v5/market/delivery-price"
+    GET_LONG_SHORT_RATIO = "/v5/market/account-ratio"
 
     def __str__(self) -> str:
         return self.value